Some refactoring, reducing the number of classes and adding the persons package
