+++ persons/Persona.java
@@ -1,16 +1,20 @@
+package person
+
 public abstract class Persona
 {
     protected String nombre;
     protected String apellido;
     protected String sexo;
     protected String DNI;
+    protected String codigo;
 
-    public Persona(String nombre, String apellido, String sexo, String DNI)
+    public Persona(String nombre, String apellido, String sexo, String DNI, String codigo)
     {
         this.nombre = nombre;
         this.apellido = apellido;
         this.sexo = sexo;
         this.DNI = DNI;
+        this.codigo = codigo;
     }
 
     public String mostrarDatos()
